VolumeDialogImpl: Update rows when adding one while the panel is showing
When a row is added while the dialog was open, for example when a remote stream
is started, the row is added as the innermost row. Due to the hidden expandable
rows we need to set the x-translation for the new row for it to not be in the
middle of the screen.

Test: 1. Start a remote stream in spotify
      2. Force close spotify
      3. Open spotify and immediately open the volume dialog by pressing a volume key
      4. Ensure the row is placed next to the default row
Change-Id: I8d5731acf85cad9b50a00c0bed116fcb1cff4f92
Signed-off-by: Fiqri Ardyansyah <fiqri191002@gmail.com>